+++ docs/changelog.rst
@@ -7,4 +7,5 @@ See :doc:`migrating to v2 <migrating_to_v2>`
 
 v2.0.1
 ~~~~~~
-Remove orjson dependency
+- Remove `orjson` dependency
+- Update ``discord.py`` requirement because this library is not compatible with ``discord.py`` from v2.4